session-validate-monir

+++ src/main/java/org/example/proflow/controller/ProfileController.java
@@ -8,6 +8,7 @@ import org.example.proflow.service.ProfileService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.example.proflow.util.Validator;
 
 import java.awt.color.ProfileDataException;
 import java.sql.SQLException;
@@ -29,7 +30,7 @@ public class ProfileController {
         this.profileService = profileService;
     }
 
-    //***LOGIN METHODS***---------------------------------------------------------------------------------------------C
+    //***LOGIN METHODS***-----------------------------------------------------------------------------------------------
     @GetMapping("/login")
     public String showLogin(){
         return "login";
@@ -77,7 +78,7 @@ public class ProfileController {
     public String addProfile(Model model) {
         Profile profile = new Profile();
         model.addAttribute("profile", profile);
-        return "signup";
+        return "signup"; //TODO har vi denne eller skal den bare hedde signup?
     }
 
     @PostMapping("/saveprofile") //PostMapping tilføjer data til database
@@ -89,6 +90,9 @@ public class ProfileController {
     //***READ PROFILE***-----------------------------------------------------------------------------------------------R
     @GetMapping("/admin") //homepage til projektleder (viser alle projekter for en projektleder)
     public String homepage(Model model, @RequestParam int profileId) throws ProfileException {
+        if(!Validator.isValid(session, profileId)) {
+            return "redirect:/homepage";
+        }
         List<Project> projectsFromProfile = profileService.getProjectsFromProfile(profileId);
         model.addAttribute("projectsFromProfile", projectsFromProfile);
         return "homepage";
@@ -103,7 +107,11 @@ public class ProfileController {
 
     //***UPDATE PROFILE***---------------------------------------------------------------------------------------------U
     @GetMapping("/edit/{profileId}")
-    public String editProfile(@PathVariable("profileId") int profileId, Model model) throws ProfileException, SQLException {
+    public String editProfile(@PathVariable("profileId") int profileId, Model model, HttpSession session)
+            throws ProfileException, SQLException {
+        if(!Validator.isValid(session, profileId)) {
+            return "redirect:/homepage";
+        }
         Profile profile = profileService.getProfileById(profileId);
         model.addAttribute("profile", profile);
         model.addAttribute("profileName", profile.getFirstName());
@@ -114,7 +122,11 @@ public class ProfileController {
     }
 
     @PostMapping("/update/{id}")
-    public String updateProfile(@PathVariable("id") int profileId, @ModelAttribute Profile profile, Model model) throws ProfileException {
+    public String updateProfile(@PathVariable("id") int profileId, @ModelAttribute Profile profile, Model model)
+            throws ProfileException {
+        if(!Validator.isValid(session, profileId)) {
+            return "redirect:/homepage";
+        }
         model.addAttribute("profile", profile);
         profile.setId(profileId);
         profileService.updateProfile(profile);
@@ -122,8 +134,11 @@ public class ProfileController {
     }
 
     //***DELETE PROFILE***---------------------------------------------------------------------------------------------D
-    @PostMapping("/{name}/remove")
+    @PostMapping("/remove/{id}")
     public String deleteProfile(@PathVariable int profileId) throws ProfileException, SQLException {
+        if(!Validator.isValid(session, profileId)) {
+            return "redirect:/homepage";
+        }
         Profile profile = profileService.getProfileById(profileId);
         profileService.deleteProfile(profile.getId());
         return "redirect:/profile";
