+++ src/main/java/org/example/proflow/controller/SubProjectController.java
@@ -1,8 +1,13 @@
 package org.example.proflow.controller;
 
+import jakarta.servlet.http.HttpSession;
+import org.example.proflow.model.Profile;
+import org.example.proflow.model.Project;
 import org.example.proflow.model.Status;
 import org.example.proflow.model.SubProject;
+import org.example.proflow.service.ProjectService;
 import org.example.proflow.service.SubProjectService;
+import org.example.proflow.util.Validator;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -13,30 +18,42 @@ import java.util.List;
 
 //TODO SubProjectController: Rette HTML sider
 //TODO SubProjectController: Rette navne/stier på endpoints
-//TODO SubProjectController: Rette exceptions til subProjectException
+//TODO SubProjectController: Rette exceptions til subProjectException ..
 
 @Controller
-@RequestMapping("homepage")
+@RequestMapping("dashboard/{profileId}/{projectId}")
 public class SubProjectController {
 
     //***ATTRIBUTES***--------------------------------------------------------------------------------------------------
+    private final ProjectService projectService;
     private final SubProjectService subProjectService;
     private Model model;
 
     //***CONSTRUCTOR***-------------------------------------------------------------------------------------------------
-    public SubProjectController(SubProjectService subProjectService) {
+    public SubProjectController(SubProjectService subProjectService, ProjectService projectService) {
         this.subProjectService = subProjectService;
+        this.projectService = projectService;
     }
 
     //***CREATE SUBPROJECT METHODS***-----------------------------------------------------------------------------------
-    @GetMapping("/{projectId}/addsubproject")
-    public String addSubProject(@PathVariable("projectId") int projectId, Model model) {
+    @GetMapping("/addsubproject")
+    public String addSubProject(@PathVariable("projectId") int projectId, Model model, HttpSession session) throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");  //Tjekker om den er logget ind
+        if(!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId); //Henter projektet fra databasen
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())){ //Tjekker om profilens ID matcher ID'et tilhørende projeketets ID
+            return "redirect:/homepage";
+        }
         model.addAttribute("projectId", projectId);
         model.addAttribute("subProject", new SubProject());
         return "homepage";
     }
 
-    @PostMapping("/{projectId}/savesubproject")
+    @PostMapping("/savesubproject")
     public String saveSubProject(@PathVariable("projectId") int projectId,
                                  @ModelAttribute("subProject") SubProject subProject, Model model) throws SQLException {
         subProject.setProjectId(projectId);
@@ -47,14 +64,35 @@ public class SubProjectController {
 
     //***READ SUBPROJECT METHODS***-------------------------------------------------------------------------------------
     @GetMapping("/subprojects")
-    public String getAllSubProjects(Model model) {
+    public String getAllSubProjects(Model model, HttpSession session, int projectId) throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");  //Tjekker om den er logget ind
+        if(!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId); //Henter projektet fra databasen
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())){ //Tjekker om profilens ID matcher ID'et tilhørende projeketets ID
+            return "redirect:/homepage";
+        }
         List<SubProject> subProjects = subProjectService.getAllSubProjects();
         model.addAttribute("SubProjects", subProjects);
         return "homepage";
     }
 
     @GetMapping("/subproject/{subprojectId}")
-    public String getSubProjectById(@PathVariable("subprojectId") int subprojectId, Model model) throws SQLException {
+    public String getSubProjectById(@PathVariable("subprojectId") int subprojectId, Model model, HttpSession session, int projectId)
+            throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");  //Tjekker om den er logget ind
+        if(!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId); //Henter projektet fra databasen
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())){ //Tjekker om profilens ID matcher ID'et tilhørende projeketets ID
+            return "redirect:/homepage";
+        }
         SubProject subProject = subProjectService.getSubProjectById(subprojectId);
         model.addAttribute("subprojectId", subprojectId);
         model.addAttribute("name", subProject.getName());
@@ -62,8 +100,18 @@ public class SubProjectController {
     }
 
     //***UPDATE SUBPROJECT METHODS***-----------------------------------------------------------------------------------
-    @GetMapping("/subproject/edit/{id}")
-    public String editSubProject(@PathVariable("id") int subProjectId, Model model) throws SQLException {
+    @GetMapping("/subproject/edit/{id}") //TODO skal den bare hedde id her?
+    public String editSubProject(@PathVariable("id") int subProjectId, Model model, HttpSession session, int projectId) throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");  //Tjekker om den er logget ind
+        if(!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId); //Henter projektet fra databasen
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())){ //Tjekker om profilens ID matcher ID'et tilhørende projeketets ID
+            return "redirect:/homepage";
+        }
         SubProject subProject = subProjectService.getSubProjectById(subProjectId);
         model.addAttribute("subProject", subProject);
         model.addAttribute("subProjectId", subProject.getId());
@@ -82,7 +130,17 @@ public class SubProjectController {
 
 
     @PostMapping("/subproject/update")
-    public String updateSubProject(@ModelAttribute SubProject subProject) throws SQLException {
+    public String updateSubProject(@ModelAttribute SubProject subProject, HttpSession session, int projectId) throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");  //Tjekker om den er logget ind
+        if(!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId); //Henter projektet fra databasen
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())){ //Tjekker om profilens ID matcher ID'et tilhørende projeketets ID
+            return "redirect:/homepage";
+        }
         int subProjectId = subProject.getId();
         subProjectService.updateSubProject(subProject);
         return "redirect:/homepage/subproject/" + subProjectId;
@@ -91,7 +149,17 @@ public class SubProjectController {
 
     //***DELETE SUBPROJECT METHODS***-----------------------------------------------------------------------------------
     @PostMapping("/subproject/delete/{subProjectId}")
-    public String deleteSubProject(@PathVariable("subProjectId") int subProjectId) throws SQLException {
+    public String deleteSubProject(@PathVariable("subProjectId") int subProjectId, int projectId, HttpSession session) throws SQLException {
+        Profile profile = (Profile) session.getAttribute("profile");  //Tjekker om den er logget ind
+        if(!Validator.isValid(session, profile.getId())) {
+            return "redirect:/homepage";
+        }
+
+        Project project = projectService.getProjectById(projectId); //Henter projektet fra databasen
+
+        if(!Validator.isProjectOwned(profile.getId(), project.getProfileId())){ //Tjekker om profilens ID matcher ID'et tilhørende projeketets ID
+            return "redirect:/homepage";
+        }
         SubProject subProject = subProjectService.getSubProjectById(subProjectId);
         subProjectService.deleteSubProject(subProjectId);
         return "rediect:/homepage/userProfile";
